chanstate: move final htlc kv helpers
Move the final-HTLC bucket helpers, byte encoding, and log-update

processing into chanstate so the KV-specific implementation lives with

the channel state package.

Keep channeldb wrappers and aliases in place so existing callers and

tests continue to compile while the remaining channel-state code moves

over in later commits.
